+++ src/main/java/io/voucherify/client/model/distribution/response/ListPublicationsResponse.java
@@ -17,7 +17,7 @@ public class ListPublicationsResponse {
 
   private String object;
 
-  private Integer total;
+  private Long total;
 
   @JsonProperty("data_ref")
   private String dataRef;
